upate

+++ app/Http/Controllers/AdmUserController.php
@@ -7,6 +7,8 @@ use App\Models\User;
 use App\Models\adm_menu;
 use App\Models\adm_role;
 use File;
+use App\Helper\menu;
+use Session;
 use Illuminate\Validation\Rules\Unique;
 
 
@@ -16,9 +18,8 @@ class AdmUserController extends Controller
     function index(){
        $data = User::all();
        $role = adm_role::all();
-       $menu = adm_menu::select('induk','kode_menu','nama_menu','route')->orderBy('kode_menu','asc')->get();
+        $menu = menu::getMenu(Session::get('role'));
         return view('page.AdmUser.index',compact('data','menu','role'));
-
     }
      function store(Request $request){
         $data = User::create([
@@ -67,11 +68,12 @@ class AdmUserController extends Controller
 
      function destroy($id){
         $data = User::where('uuid',$id)->first();
+        File::delete('assets/img/user/'.$data->foto);
         $data->delete();
         if($data){
-            return redirect()->back()->with('success','Menu Behasil dihapus');
+            return redirect()->back()->with('success','User Behasil dihapus');
         }else{
-            return redirect()->back()->with('fail','Menu Gagal dihapus');
+            return redirect()->back()->with('fail','User Gagal dihapus');
         }
     }
 
@@ -91,14 +93,17 @@ class AdmUserController extends Controller
         }
     }
     function updatePassword($id, Request $request){
-        $data = User::where('uuid',$id)->update([
-            'password' => md5($request->password),
-           
-        ]);
+
+        $data = User::where('uuid',$id)->first();
+        if(md5($request->current_password) == $data->password ){
+            $data->update([
+                'password' => md5($request->new_password)
+            ]);
+        }
         if($data){
-            return redirect()->back()->with('success','Menu Behasil dihapus');
+            return redirect()->back()->with('success','Sandi Behasil diubah');
         }else{
-            return redirect()->back()->with('fail','Menu Gagal dihapus');
+            return redirect()->back()->with('fail','gagal Gagal diubah');
         }
     }
     public function updatePhoto($id, Request $request)
@@ -111,18 +116,19 @@ class AdmUserController extends Controller
         // Mengambil file photo dari request
         $photo = $request->file('photo');
 
-        $imagename=uniqid().'.'.$photo->getClientOriginalExtension();
+        $imagename = uniqid().'.'.$photo->getClientOriginalExtension();
 
         // Menyimpan file photo ke folder public/photos (pastikan folder sudah ada dan writable)
         $photoPath = $photo->move(public_path('assets/img/user'),$imagename);
 
         // Mengupdate data user dengan informasi foto baru
         $user = User::where('uuid', $id)->first();
-        if (file_exists(public_path('assets/img/user/'.$user->foto))){
+        if (!is_int(strpos($user->foto, 'profile_small')) && file_exists(public_path('assets/img/user/'.$user->foto))){
             File::delete('assets/img/user/'.$user->foto);
         }
         $user->foto = $imagename;
         $user->save();
+        Session::put('photo',$imagename);
 
         return redirect()->back()->with('success', 'Photo updated successfully!');
     }
